contrl update
